backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
django-filter==25.1
django-widget-tweaks==1.5.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
